+++ docs/changelog.rst
@@ -19,6 +19,7 @@ Changes in 0.8
 - Added _instance to EmbeddedDocuments pointing to the parent (MongoEngine/mongoengine#139)
 - Inheritance is off by default (MongoEngine/mongoengine#122)
 - Remove _types and just use _cls for inheritance (MongoEngine/mongoengine#148)
+- Only allow QNode instances to be passed as query objects (MongoEngine/mongoengine#199)
 
 Changes in 0.7.9
 ================